// frontend/src/api/models/kanbanBoardSummaryDtoDTO.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API
 * Clean Architecture API with JWT Authentication
 * OpenAPI spec version: v1
 */

export interface KanbanBoardSummaryDtoDTO {
  id?: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt?: string;
  updatedAt?: string;
  taskCount?: number;
  completedTaskCount?: number;
}